requirements.txt
python-fasthtml>=0.12.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.4.0

# LLM + agents
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
openai>=1.40.0
tiktoken>=0.7.0

# Database
psycopg[binary,pool]>=3.2.0
sqlalchemy>=2.0.30
pgvector>=0.3.0

# Local embeddings (no API key required — ONNX runtime, ~100 MB model)
fastembed>=0.4.0

# Data generation / utils
faker>=26.0.0
numpy>=2.0.0
pandas>=2.2.0
python-dateutil>=2.9.0
pyyaml>=6.0.2

# Analytics (NL → SQL → Plotly) + optional web search for sourcing agents
plotly>=6.0.0
tavily-python>=0.7.0
exa-py>=2.0.0

# Dev / testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# Demo artifact generation (screenshots → gif + pdf)
pillow>=10.0.0
reportlab>=4.0.0
playwright>=1.45.0
